+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [2.3.1.2.6.4] - 2023-10-16
 
+### Added
+
+- Weapon classes (`carbine`, `polearm`, `sidearm`...).
+
 ### Changes
 
 - Art Updates.